requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
opencv-python
Pillow
spacy
serpapi
numpy
scikit-learn
python-dotenv
